Adding conditional for plunkr
Proper UX for this example indicates the Home should be selected on initial load. Currently this does not work because the initial routing event has the id of the plunkr page instead of "" or "/". This is a weak attempt to fix it. Please let me know a better way to achieve this and I will implement it and reverse this commit.